Initial commit for integration of report sale with backend
